requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
configparser==3.5.0
idna==2.5
praw==5.0.1
prawcore==0.11.0
python-dotenv==0.6.4
pytz==2017.2
redis==2.10.5
requests==2.18.1
update-checker==0.16
urllib3==1.21.1
